Split table.js into what it actually is
713 lines and six responsibilities become five modules, none over 330 lines.
madge circular dependencies: **1 → 0**. The survivor was `ExpandToggle ⇄ table`,
and it went the moment `ExpandToggle` started importing the axis and layout
modules directly.

| Module | Lines | Holds |
|---|---|---|
| `components/table.js` | 151 | the scaffold, and only that |
| `rendering/axes.js` | 330 | `renderAxes` + its 8 private helpers |
| `components/svgLayout.js` | 140 | `updateSVGLayout`, `applyZoomTransform` |
| `components/spectrogramImage.js` | 85 | `setupSpectrogramImage` |
| `utils/coordinates.js` | +2 exports | `getRenderDimensions`, `calculateVisibleDataRange` |

`components/table.js` is now imported by exactly one module, `DOMSetup.js`,
which is what a scaffold should look like. `rendering/axes.js` finally exists
beside `cursors.js` and `symbols.js`, closing part of GF-38.

**Two placements differ from the plan, and the reason is the plan's own
SC-002.** Putting `getRenderDimensions` in `spectrogramImage.js` and
`calculateVisibleDataRange` + `applyZoomTransform` in `viewport.js` — as
tasks.md specifies — creates two *new* cycles: `spectrogramImage ⇄ svgLayout`
(both need render dimensions) and `viewport ⇄ svgLayout` (`updateSVGLayout`
ends by applying the zoom transform). Trading one cycle for two is not a split.

So the two pure helpers went to `utils/coordinates.js`, which imports nothing
and already had `renderSize` as a private duplicate of `getRenderDimensions`
— now the one implementation — and `applyZoomTransform` went beside
`updateSVGLayout`, which is what calls it. `core/viewport.js` remains the
single home for zoom *math* (`zoomIn`/`zoomOut`/`zoomReset`/`setZoom`/
`zoomAtImagePoint`/`panByNormalized`); what moved out of it is transform
*application*, so FR-007 and AS-3.4 stand. Recorded for ADR-018.

The two relocated helpers take `(viewport, element)` like every other export
in `coordinates.js`, rather than `(instance)`. That is the only signature
change; the bodies are verbatim.

Strict errors 491 → 469. 21 of those came from typing `state` on the class —
`instance.state` had been inferred as possibly-undefined, which the moved call
sites surfaced by passing it where a `Viewport` was expected. The one
`TS2564` it leaves behind is the constructor's legacy-browser early return,
resolved with the explicit-initialization work.

Reach-ins 237 → 229: the new modules bind the viewport once per function
instead of walking `instance.state` on each line.

Suite: 243 passed, **with no spec file edited** — AS-3.1's actual gate.
