+++ hygiene-baseline.json
@@ -1,11 +1,11 @@
 {
   "_comment": "Debt ratchet baselines (specs/164-quality-ratchets). `yarn hygiene` fails any change that raises a count above its baseline. When your change lowers a count, lower the baseline here in the same PR so the improvement is locked in.",
-  "circularDependencies": 1,
+  "circularDependencies": 0,
   "unusedExportModules": 5,
   "waitForTimeoutOccurrences": 1,
   "_comment_strictTypeErrors": "Temporary (specs/167-structural-refactor Story 1). Errors reported by tsc under tsconfig.strict.json, the overlay enabling noImplicitAny, strictNullChecks and strictPropertyInitialization. Burned to zero, at which point the flags move into tsconfig.json and both the overlay and this entry are deleted.",
-  "strictTypeErrors": 491,
+  "strictTypeErrors": 469,
   "_comment_instanceSurface": "specs/167-structural-refactor Story 5. instanceStateReachIns counts lines containing `instance.state` under src/; instanceFields counts class-field declarations between `export class GramFrame` and its constructor in src/main.js.",
-  "instanceStateReachIns": 237,
+  "instanceStateReachIns": 229,
   "instanceFields": 56
 }
