+++ specs/167-structural-refactor/tasks.md
@@ -161,11 +161,11 @@ import rewiring; `git diff -M` must show renames. Any behaviour change is a
 separate task. AS-3.1 requires zero behavioural diff — a spec that needed editing
 means the move was not a move.
 
-- [ ] T023 [US3] Create `src/rendering/axes.js`: move `renderAxes` (`table.js:326-358`) and the 8 private helpers (`table.js:412-687` — `renderTimeAxis`, `renderFrequencyAxis`, `calculateAxisTicks`, `formatFrequencyLabels`, `renderAxisLine`, `renderAxisTicks`, `renderAxisLabels`) verbatim; export only `renderAxes`; rewire `src/core/viewport.js` and `src/components/ExpandToggle.js` — see [contracts/axes.md](./contracts/axes.md)
-- [ ] T024 [US3] Create `src/components/spectrogramImage.js` (`setupSpectrogramImage` from `table.js:140-205`, `getRenderDimensions` from `table.js:22-34`) and `src/components/svgLayout.js` (`updateSVGLayout` from `table.js:206-265`); rewire `src/core/initialization/UISetup.js`, `src/components/ExpandToggle.js`, `src/modes/harmonics/HarmonicsMode.js` and `src/core/viewport.js`
-- [ ] T025 [US3] Move `applyZoomTransform` (`table.js:266-325`) and `calculateVisibleDataRange` (`table.js:359-411`) into `src/core/viewport.js`, giving zoom math one home for wheel, keyboard, PanMode and API callers (FR-007, AS-3.4); rewire `src/modes/harmonics/HarmonicsMode.js` and `src/modes/harmonics/ManualHarmonicModal.js`
-- [ ] T026 [US3] Verify `src/components/table.js` is scaffold-only (~135 lines: `setupComponentTable`, private `createComponentStructure` and `replaceConfigTable`) and imported by exactly one module, `src/core/initialization/DOMSetup.js`; confirm all five modules are under 350 lines
-- [ ] T027 [US3] Lower `circularDependencies` in `hygiene-baseline.json` to 0 and confirm no new cycle appeared (AS-3.3, SC-002)
+- [x] T023 [US3] Create `src/rendering/axes.js`: move `renderAxes` (`table.js:326-358`) and the 8 private helpers (`table.js:412-687` — `renderTimeAxis`, `renderFrequencyAxis`, `calculateAxisTicks`, `formatFrequencyLabels`, `renderAxisLine`, `renderAxisTicks`, `renderAxisLabels`) verbatim; export only `renderAxes`; rewire `src/core/viewport.js` and `src/components/ExpandToggle.js` — see [contracts/axes.md](./contracts/axes.md)
+- [x] T024 [US3] Create `src/components/spectrogramImage.js` (`setupSpectrogramImage` from `table.js:140-205`, `getRenderDimensions` from `table.js:22-34`) and `src/components/svgLayout.js` (`updateSVGLayout` from `table.js:206-265`); rewire `src/core/initialization/UISetup.js`, `src/components/ExpandToggle.js`, `src/modes/harmonics/HarmonicsMode.js` and `src/core/viewport.js`
+- [x] T025 [US3] Move `applyZoomTransform` (`table.js:266-325`) and `calculateVisibleDataRange` (`table.js:359-411`) into `src/core/viewport.js`, giving zoom math one home for wheel, keyboard, PanMode and API callers (FR-007, AS-3.4); rewire `src/modes/harmonics/HarmonicsMode.js` and `src/modes/harmonics/ManualHarmonicModal.js`
+- [x] T026 [US3] Verify `src/components/table.js` is scaffold-only (~135 lines: `setupComponentTable`, private `createComponentStructure` and `replaceConfigTable`) and imported by exactly one module, `src/core/initialization/DOMSetup.js`; confirm all five modules are under 350 lines
+- [x] T027 [US3] Lower `circularDependencies` in `hygiene-baseline.json` to 0 and confirm no new cycle appeared (AS-3.3, SC-002)
 
 **Checkpoint**: SC-002 met at 0. `rendering/axes.js` exists, closing part of GF-38.
