+++ src/components/spectrogramImage.js
@@ -1,0 +1,86 @@
+/**
+ * Spectrogram image setup and scaling.
+ *
+ * Loads the configured image, records its natural (post-downscale) dimensions,
+ * seeds the render dimensions from them, and triggers the first layout, axis
+ * render and expand-toggle mount. Split out of `components/table.js`
+ * (spec 167, FR-004).
+ */
+
+/// <reference path="../types.js" />
+
+import { dispatch } from '../core/state.js'
+import { renderAxes } from '../rendering/axes.js'
+import { createExpandToggle } from './ExpandToggle.js'
+import { updateSVGLayout } from './svgLayout.js'
+
+// Maximum image width in pixels - images wider than this will be scaled down
+const MAX_IMAGE_WIDTH = 1200
+
+/**
+ * Set up spectrogram image display within SVG container
+ * @param {GramFrame} instance - GramFrame instance
+ * @param {string} imageUrl - URL of the spectrogram image
+ */
+export function setupSpectrogramImage(instance, imageUrl) {
+  if (!instance.spectrogramImage || !imageUrl) {
+    return
+  }
+  
+  // Set image source
+  instance.spectrogramImage.setAttributeNS('http://www.w3.org/1999/xlink', 'href', imageUrl)
+  
+  // Store URL in state
+  instance.state.imageDetails.url = imageUrl
+  
+  // Load image to get natural dimensions
+  const tempImg = new Image()
+  tempImg.onload = function() {
+    // Dimensions are known, so the panel is about to render for real
+    instance.container.classList.remove('gram-frame-loading')
+
+    // Get original dimensions
+    let imageWidth = tempImg.naturalWidth
+    let imageHeight = tempImg.naturalHeight
+    
+    // Apply automatic scaling for images wider than the maximum allowed width
+    if (imageWidth > MAX_IMAGE_WIDTH) {
+      const scaleFactor = MAX_IMAGE_WIDTH / imageWidth
+      imageWidth = MAX_IMAGE_WIDTH
+      imageHeight = Math.round(imageHeight * scaleFactor)
+      
+      console.log(`GramFrame: Scaling down large image from ${tempImg.naturalWidth}x${tempImg.naturalHeight} to ${imageWidth}x${imageHeight} (scale factor: ${scaleFactor.toFixed(3)})`)
+    }
+    
+    // Store scaled dimensions as natural dimensions
+    const imageDetails = instance.state.imageDetails
+    imageDetails.naturalWidth = imageWidth
+    imageDetails.naturalHeight = imageHeight
+
+    // Initialise render dimensions to natural dimensions. Expand updates these
+    // to fill available space; collapse restores them to natural exactly.
+    imageDetails.renderWidth = imageWidth
+    imageDetails.renderHeight = imageHeight
+
+    // Update SVG layout
+    updateSVGLayout(instance)
+
+    // Render axes
+    renderAxes(instance)
+
+    // Mount the expand toggle now that natural dimensions (and thus the
+    // landscape test) are known. No-op for portrait/square images.
+    createExpandToggle(instance)
+
+    // Notify listeners of updated dimensions
+    dispatch(instance)
+  }
+  tempImg.onerror = function() {
+    // Without dimensions nothing can be rendered, so replace the loading
+    // caption with a failure one instead of leaving it spinning forever
+    console.error(`GramFrame: Failed to load spectrogram image: ${imageUrl}`)
+    instance.container.classList.remove('gram-frame-loading')
+    instance.container.classList.add('gram-frame-image-error')
+  }
+  tempImg.src = imageUrl
+}
